combine PackageInfo into Manifest
