+++ crates/notion-core/src/serial/mod.rs
@@ -4,7 +4,6 @@ pub mod catalog;
 pub mod config;
 pub mod index;
 pub mod manifest;
-pub mod package;
 pub mod plugin;
 pub mod version;
